# server/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type MaintenanceRecord struct {
	ID          pgtype.UUID        `json:"id"`
	VehicleID   pgtype.UUID        `json:"vehicle_id"`
	Date        pgtype.Date        `json:"date"`
	Description string             `json:"description"`
	Mileage     int32              `json:"mileage"`
	Cost        *string            `json:"cost"`
	Notes       *string            `json:"notes"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Vehicle struct {
	ID        pgtype.UUID        `json:"id"`
	Make      string             `json:"make"`
	Model     string             `json:"model"`
	Year      int32              `json:"year"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type VehicleImage struct {
	ID        int32       `json:"id"`
	VehicleID pgtype.UUID `json:"vehicle_id"`
	Filename  string      `json:"filename"`
}